Fix Klarna  redirect flow for declined payments
If a customer completes the Klarna payment flow and the payment is
declined, Klarna assumes that the customer can choose another payment
method offered by Klarna e.g. credit card. The plugin  is currently limited
to pay later and results in order creation even though the payment failed.
To prevent this a check is now in place which cancles the order creation
in that case, so the customer can choose an alternative.

Signed-off-by: Tim Marx <tim.marx2@gmail.com>